refactor: 컴포넌트, 클래스들의 역할과 책임에 맞는 패키지 명으로 수정
리팩토링 내용
- 기존의 vehicle -> car로 패키지 명 변경
- 모호한 wrapper 패키지 제거
