+++ src/test/java/racingcar/controller/racing/wrapper/LapsTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.controller.Laps;
 
 public class LapsTest {
