+++ src/test/java/racingcar/model/car/CarNameTest.java
@@ -1,4 +1,4 @@
-package racingcar.model.racingvehicle.wrapper;
+package racingcar.model.car;
 
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.model.car.CarName;
 
 class CarNameTest {
