+++ src/test/java/racingcar/model/car/ForwardDistanceTest.java
@@ -1,10 +1,9 @@
-package racingcar.model.engine.wrapper;
+package racingcar.model.car;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.model.engine.wrapper.ForwardDistance;
 
 public class ForwardDistanceTest {
 
@@ -13,7 +12,7 @@ public class ForwardDistanceTest {
     void 전진_안한경우_DISTANCE_길이는_0() {
         ForwardDistance distance = new ForwardDistance();
         forwardForCount(distance, 0);
-        assertThat(String.valueOf(distance)).isEqualTo("");
+        assertThat(distance.distance()).isEqualTo("");
 
     }
 
@@ -22,7 +21,7 @@ public class ForwardDistanceTest {
     void 두_번_전진시_DISTANCE_길이는_2() {
         ForwardDistance distance = new ForwardDistance();
         forwardForCount(distance, 2);
-        assertThat(String.valueOf(distance)).isEqualTo("--");
+        assertThat(distance.distance()).isEqualTo("--");
     }
 
     @Test
@@ -30,12 +29,12 @@ public class ForwardDistanceTest {
     void 다섯_번_전진시_DISTANCE_길이는_5() {
         ForwardDistance distance = new ForwardDistance();
         forwardForCount(distance, 5);
-        assertThat(String.valueOf(distance)).isEqualTo("-----");
+        assertThat(distance.distance()).isEqualTo("-----");
     }
 
     private void forwardForCount(ForwardDistance distance, int count) {
         for (int i = 0; i< count; i++) {
-            distance.increase();
+            distance.hitFrom(new Rpm(5));
         }
     }
 }
